requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
backoff==2.2.1
branca==0.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==3.2.23
django-redis==5.4.0
djangorestframework==3.12.4
drf-yasg==1.21.9
folium==0.19.5
frozenlist==1.5.0
geographiclib==2.0
geopy==2.4.1
idna==3.10
inflection==0.5.1
Jinja2==3.1.5
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.0.2
opencage==3.0.4
packaging==24.2
pandas==2.2.3
propcache==0.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
xyzservices==2025.1.0
yarl==1.18.3
